requirements.txt
# --- Core Framework (H100 Optimized) ---
# Note: torch is handled in setup script, but listed here for reference
# torch>=2.2.0
# torchvision>=0.17.0

# --- Model Architectures & Layers ---
timm>=0.9.12              # ViT/ConvNeXt implementations, highly optimized
einops>=0.7.0             # Tensor manipulation for ViTs

# --- Data Processing ---
numpy>=1.26.0             # Core numerical computing
pandas>=2.1.0             # Data analysis and CSV handling
Pillow>=10.2.0            # Image loading
kornia>=0.7.1             # Differentiable image transforms (GPU-accelerated)
scikit-learn>=1.4.0       # t-SNE, metrics, preprocessing

# --- Configuration & CLI ---
PyYAML>=6.0.1             # Config parsing
click>=8.1.7              # CLI framework (for future extensibility)
typing_extensions>=4.9.0  # Backported typing features

# --- Math & Analysis ---
scipy>=1.12.0             # CKA computation, scientific computing
statsmodels>=0.14.0       # Statistical tests for significance

# --- Visualization ---
matplotlib>=3.8.0         # Plotting
seaborn>=0.13.0           # Statistical visualization
rich>=13.7.0              # Pretty terminal output

# --- Experiment Tracking ---
wandb>=0.16.2             # Weights & Biases integration
tqdm>=4.66.0              # Progress bars

# --- Development & Testing ---
pytest>=8.0.0             # Testing framework
black>=24.1.0             # Code formatting
isort>=5.13.0             # Import sorting
flake8>=7.0.0             # Linting
mypy>=1.8.0               # Type checking

# --- Hardware Acceleration Helpers ---
ninja>=1.11.1             # Fast compilation for flash-attn
psutil>=5.9.8             # System monitoring

# --- Optional (installed separately in setup script) ---
# flash-attn>=2.5.0       # Flash Attention 2 (requires compilation)
# pyhessian                # Hessian analysis (installed from git)
